// frontend/src/api/models/handlerCreateOverrideHTTPRequest.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { DtoFlagValueDTO } from './dtoFlagValueDTO'
import type { HandlerCreateOverrideHTTPRequestTargetType } from './handlerCreateOverrideHTTPRequestTargetType'

/**
 * Request body for creating a flag override
 */
export interface HandlerCreateOverrideHTTPRequest {
  expires_at?: string
  /** @maxLength 500 */
  reason?: string
  target_id: string
  target_type: HandlerCreateOverrideHTTPRequestTargetType
  value?: DtoFlagValueDTO
}