// frontend/src/api/models/handlerCreateOverrideHTTPRequestTargetType.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerCreateOverrideHTTPRequestTargetType =
  (typeof HandlerCreateOverrideHTTPRequestTargetType)[keyof typeof HandlerCreateOverrideHTTPRequestTargetType]

export const HandlerCreateOverrideHTTPRequestTargetType = {
  user: 'user',
  tenant: 'tenant',
} as const