// frontend/src/api/models/dtoOverrideResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { DtoFlagValueDTO } from './dtoFlagValueDTO'

export interface DtoOverrideResponse {
  created_at?: string
  created_by?: string
  expires_at?: string
  flag_key?: string
  id?: string
  is_expired?: boolean
  reason?: string
  target_id?: string
  target_type?: string
  updated_at?: string
  value?: DtoFlagValueDTO
}